# specs/apis/dist/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your Product's API",
    "version": "1.0.0",
    "description": "The API used internally by web clients."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate User",
        "operationId": "loginUser",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout User",
        "operationId": "logoutUser",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful logout"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get Users",
        "operationId": "getUsers",
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/user"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create User",
        "operationId": "createUser",
        "tags": [
          "users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          },
          "409": {
            "description": "User already exists"
          }
        }
      }
    },
    "/todos": {
      "get": {
        "summary": "List all todos",
        "tags": [
          "todos"
        ],
        "responses": {
          "200": {
            "description": "A list of todo items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/todo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new todo",
        "tags": [
          "todos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created todo item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/todo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
    "/todos/{id}": {
      "put": {
        "summary": "Update a todo",
        "tags": [
          "todos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated todo item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/todo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Todo item not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a todo",
        "tags": [
          "todos"
        ],
        "responses": {
          "204": {
            "description": "Todo item successfully deleted"
          },
          "404": {
            "description": "Todo item not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "$ref": "#/components/schemas/user"
      },
      "Todo": {
        "$ref": "#/components/schemas/todo"
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "user": {
        "type": "object",
        "required": [
          "id",
          "email",
          "name"
        ],
        "properties": {
          "id": {
            "type": "number",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "email",
          "name"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "todo": {
        "type": "object",
        "required": [
          "id",
          "title",
          "completed",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the todo item"
          },
          "title": {
            "type": "string",
            "description": "The title/description of the todo item"
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the todo item has been completed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the todo item was created"
          }
        }
      },
      "CreateTodoRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "The title/description of the todo item"
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "required": [
          "title",
          "completed"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "The title/description of the todo item"
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the todo item has been completed"
          }
        }
      }
    }
  }
}